requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Model loading
transformers>=4.30.0
timm>=0.9.0

# Performance
numba>=0.57.0

# Data handling
pillow>=9.5.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Analysis (optional)
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
